refactor: move the InfluxDB setup to config and shared helpers to common #307
