+++ backend/src/main/java/com/occupi/feature/forecast/ForecastServiceImpl.java
@@ -2,9 +2,9 @@ package com.occupi.feature.forecast;
 
 import com.influxdb.v3.client.InfluxDBClient;
 import com.influxdb.v3.client.query.QueryOptions;
+import com.occupi.common.InfluxTime;
+import com.occupi.common.TimeSlots;
 import com.occupi.config.CacheConfig;
-import com.occupi.feature.chart.TimeSlots;
-import com.occupi.feature.database.InfluxTime;
 import com.occupi.feature.forecast.dto.ForecastPoint;
 import com.occupi.feature.forecast.dto.ForecastResponse;
 import lombok.extern.slf4j.Slf4j;
