+++ backend/src/test/java/com/occupi/feature/forecast/ForecastInfluxIntegrationTest.java
@@ -1,7 +1,7 @@
 package com.occupi.feature.forecast;
 
 import com.influxdb.v3.client.InfluxDBClient;
-import com.occupi.feature.chart.TimeSlots;
+import com.occupi.common.TimeSlots;
 import com.occupi.feature.database.model.OccupancyData;
 import com.occupi.feature.database.repository.OccupancyRepository;
 import com.occupi.feature.forecast.dto.ForecastPoint;
